% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_lizard_data.R
\name{import_lizard_data}
\alias{import_lizard_data}
\title{Import lizard data
This function imports a CSV file of processed field data of lizard individuals
of different species including sites, their location (coordinates), morphology data
and reflectance means}
\usage{
import_lizard_data(path, species = "")
}
\description{
Import lizard data
This function imports a CSV file of processed field data of lizard individuals
of different species including sites, their location (coordinates), morphology data
and reflectance means
}
