% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_estimate_deb.R
\name{m_estimate_deb}
\alias{m_estimate_deb}
\title{m_estimate_deb}
\usage{
m_estimate_deb(param)
}
\arguments{
\item{param}{List of input parameters for ectotherm function.}
}
\description{
This function uses the estimation function of the Dynamic Energy Budget model
to calculate the input parameters for DEB.
}
