% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_ectotherm_input.R
\name{m_extract_ectotherm_input}
\alias{m_extract_ectotherm_input}
\title{m_extract_ectotherm_input}
\usage{
m_extract_ectotherm_input()
}
\description{
This function extracts input data for the ectotherm function of NicheMapR
from our data
}
