% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_microclim.R
\name{m_extract_microclim}
\alias{m_extract_microclim}
\title{Extract microclimate
This function extracts the output data of the micro_global() function of NicheMapR
based on our data structure (see example_lizard_data.csv in the testthat folder in tests)}
\usage{
m_extract_microclim(file, species = "")
}
\arguments{
\item{file}{Data file in csv format (set working directory to folder with all data files)}

\item{species}{Species of interest in format "Genus_species"}
}
\value{
List of vectors of micro climate data for each location
}
\description{
Extract microclimate
This function extracts the output data of the micro_global() function of NicheMapR
based on our data structure (see example_lizard_data.csv in the testthat folder in tests)
}
