% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_microclim_input.R
\name{m_extract_microclim_input}
\alias{m_extract_microclim_input}
\title{Extract microclimate input}
\usage{
m_extract_microclim_input(locations)
}
\arguments{
\item{locations}{An R factor containing the location IDs of interest (as in example_lizard_data.csv)}
}
\value{
List of vectors of input data for each location for the micro_global function
}
\description{
This function extracts the input data of the micro_global() function of NicheMapR
based on our data structure (see example_lizard_data.csv in the testthat folder in tests)
}
