% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_get_microclim.R
\name{m_get_microclim}
\alias{m_get_microclim}
\title{m_get_microclim}
\arguments{
\item{loc_list}{A list containing information on location (location ID, latitude & longitude,
nature of soil, soil reflectance, soilgrids (downloading: yes or no))}
}
\value{
list of locations (as sublists) with corresponding microclimate data
}
\description{
This function models the microclimate of locations in the given
input list (= output of m_extract_microclim_input)
}
