% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_activity.R
\name{m_plot_activity}
\alias{m_plot_activity}
\title{Plot activity times over the organisms reflectance}
\usage{
m_plot_activity(multi_ecto, save_plot = FALSE)
}
\arguments{
\item{multi_ecto}{A tidy data frame of summarised output results of the ecotherm function
containing the activity times (total, percentage, ratios, etc.), absorptivity, weight,
yearly averaged microclimate variables per scenario and location (for details see ?m_tidy_output).}

\item{save_plot}{Boolean whether the plot should be saved or not (default = FALSE).}
}
\value{
Plot
}
\description{
This function plots the output of the biophysical model of ectotherms
from the NicheMapR package.
}
