% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_micro.R
\name{m_plot_micro}
\alias{m_plot_micro}
\title{Plot microclimate conditions}
\usage{
m_plot_micro(ectoall, save_plot = FALSE)
}
\arguments{
\item{ectoall}{A climate scenario which is a list of lists
of locations which are lists with the structure of the output of the micro_global()
function (see vignette of microclimate in NicheMapR package)}

\item{save_plot}{Boolean whether the microclimate plot should be saved or not
(default = FALSE)}
}
\value{
Plot
}
\description{
This function plots the microclimate conditions from the output of
multiple ectotherm models of different conditions (locations, times, emission
scenarios) together in one plot
}
