% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_run_biophysical.R
\name{m_run_biophysical}
\alias{m_run_biophysical}
\title{m_run_biophysical}
\usage{
m_run_biophysical(liz_file = "example_lizard_data.csv",
  species = "Karusasaurus_polyzonus",
  loc_file = "example_coordinates.csv", nyears = 1, ndays = 12,
  burrow = FALSE, DEB = FALSE, timeper = "present", rcp = "none")
}
\arguments{
\item{liz_file}{A file with lizard data (see example_lizard_data.csv) containing at least
one individual with location, morphology and reflectance data.}

\item{species}{The species to be looked at, which will be selected from the data frame and
split into sub groups of morphs if present.}

\item{loc_file}{File with information on locations (see example_coordinates.csv).}

\item{nyears}{Number of years the model is run.}

\item{ndays}{Number of days modeled per year.}

\item{burrow}{Boolean whether lizard is allowed to seek shelter in burrow.}

\item{DEB}{Boolean stating wheather the ectotherm should be run with or without
the Dynamic Energy Budget model.}

\item{timeper}{Character string of the time period for the climate data /
predictions ("present", "2040_2059", "2080_2099").}

\item{rcp}{Character string to specify the emission scenario in case of climate
predictions ("none", "45", "85").}
}
\value{
List of lists with output if ectotherm function for each species/population/entity
}
\description{
This function runs a full analysis of the climate niche starting with importing
our standard dataframe, running microclimate model and ectotherm function subsequently.
}
