% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_run_climate_analysis.R
\name{m_run_climate_analysis}
\alias{m_run_climate_analysis}
\title{m_run_climate_analysis}
\usage{
m_run_climate_analysis(liz_file = "example_lizard_data.csv",
  species = "", loc_file = "example_coordinates.csv")
}
\arguments{
\item{liz_file}{A file with lizard data (see example_lizard_data.csv) containing at least
one individual with location, morphology and reflectance data.}

\item{species}{The species to be looked at, which will be selected from the data frame and
split into sub groups of morphs if present.}

\item{loc_file}{File with information on locations (see example_coordinates.csv).}
}
\value{
List of lists with output if ectotherm function for each species/population/entity
}
\description{
This function runs a full analysis of the climate niche starting with importing
our standard dataframe, running microclimate model and ectotherm function subsequently.
}
