% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_tidy_output_ind.R
\name{m_tidy_output_ind}
\alias{m_tidy_output_ind}
\title{Tidy(verse) up NMR results - individual input data}
\usage{
m_tidy_output_ind(multi_all, monthly_clim = FALSE, avg_loc_out = FALSE)
}
\arguments{
\item{multi_all}{A list of different climate scenarios which are lists
of locations which are lists containing the microclimate and biophysical
model outputs plus some parameter settings (see vignette of 'ectotherm'
and 'micro_global' in NicheMapR package).}

\item{monthly_clim}{Boolean to define whether the microclimate variables are
reported per month or as yearly average (default = FALSE).}

\item{avg_loc_out}{When the input data is per individual, this parameter decides
if the output should be averaged (TRUE) or not (FALSE).}
}
\value{
data.frame
}
\description{
This function tidys up the output of the biophysical model of ectotherms
from the NicheMapR package and returns it as one dataframe. This function takes over
the individual input data tidying.
}
