% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_tidy_output_loc.R
\name{m_tidy_output_loc}
\alias{m_tidy_output_loc}
\title{Tidy(verse) up NMR results - location input data}
\usage{
m_tidy_output_loc(multi_all, monthly_clim = FALSE)
}
\arguments{
\item{multi_all}{A list of different climate scenarios which are lists
of locations which are lists containing the microclimate and biophysical
model outputs plus some parameter settings (see vignette of 'ectotherm'
and 'micro_global' in NicheMapR package).}

\item{monthly_clim}{Boolean to define whether the microclimate variables are
reported per month or as yearly average (default = FALSE).}
}
\value{
data.frame
}
\description{
This function tidys up the output of the biophysical model of ectotherms
from the NicheMapR package and returns it as one dataframe. This function takes over
the tidying of input data that is averaged over location.
}
