% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_whole_climate_analysis.R
\name{m_whole_climate_analysis}
\alias{m_whole_climate_analysis}
\title{Whole climate analysis}
\usage{
m_whole_climate_analysis(
  times = c("present"),
  rcps = c("none"),
  data_dir = "/.",
  liz_file = "Darkness_Morpho_Data_Adjusted.csv",
  species,
  loc_file = "Coordinates_Clean_notransect_MOK.csv",
  loc_mean = FALSE,
  physio_file = "Physio_info.csv",
  nyears = 1,
  ndays = 12,
  burrow = TRUE,
  burrowtype = "sunshade",
  burrowdepth = 9,
  DEB = FALSE,
  shade = c(0, 50),
  plot = FALSE,
  save_plot = FALSE,
  monthly = FALSE
)
}
\arguments{
\item{times}{Vector of strings listing the different time periods to run the model on.
Be aware that you need to have downloaded the according climate data and saved it in
the appropriate folders. Default is the time period of 1961 to 1990 ("present") used
in New et al. (2002), the standard climate dataset used in NicheMapR (Kearney & Porter,
2017). Options: "present", "2040_2059", "2080_2099", "presentCCKP" to compare
the 1961-1990 of CCKP with the one from NicheMapR, and "presentNASA" to use more recent
(1983-2017) climate data.}

\item{rcps}{Vector of strings listing the Representative Concentration Pathway (RCP)
scenarios of greenhouse gas concentrations. "45" and "85" stand for RCP4.5 and RCP8.5,
respectively.}

\item{liz_file}{A file with lizard data (see example_lizard_data.csv) containing at least
one individual with location, morphology and reflectance data.}

\item{species}{The species to be looked at, which will be selected from the data frame and
split into sub groups of morphs if present.}

\item{loc_file}{File with information on locations (see example_coordinates.csv).}

\item{loc_mean}{Boolean whether the biophysical model should be run with location means
to decrease computational demand (loc_mean = TRUE) or if it should be run per individual (FALSE).}

\item{physio_file}{File path with physiological data of the lizards per location (see
'example_lizard_data.csv' for structure of dataframe).}

\item{nyears}{Number of years the model is run.}

\item{ndays}{Number of days modeled per year.}

\item{burrow}{Boolean whether lizard is allowed to seek shelter in burrow.}

\item{burrowtype}{Character string defining the burrow type: "sun" (always in the sun),
"shade" (always in the shade), "sunshade" (organism decides if burrow is in the sun
or in the shade).}

\item{burrowdepth}{The burrows depth in "soil nodes", a numeric ranging from 2 to 10.
2 equals to 2.5 cm, 3 to 5 cm, 4 to 10 cm, 5 to 15 cm, 6 to 20 cm, 7 to 30 cm, 8 to 50 cm,
9 to 100 cm and 10 to 200 cm. These are also the steps the lizard takes while burrowing.}

\item{DEB}{Boolean stating wheather the ectotherm should be run with or without
the Dynamic Energy Budget model. !!!NOT FULLY IMPLEMENTED AND TESTED YET!!!}

\item{shade}{Vector of 2 numerics defining minimum and maximum shade. Default is 0 %
minimum shade, 50 % maximum shade.}

\item{plot}{Boolean whether to produce plots of activity pattern of individual
lizards or not.}

\item{save_plot}{Boolean whether plots shall be saved or not.}

\item{monthly}{Boolean whether the final dataframe should be organised per individual
or per month at each location.}
}
\value{
Dataframe with the collective physiological and activity data of all
individuals or populations and their respective climatic environments.
}
\description{
All-encompassing function to run all different scenarios at once
(present & futures), given the input
}
