% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/k.poly._analysis_script.R
\docType{data}
\name{path}
\alias{path}
\title{General analysis of lizard data set
This is a script to check the data of Karusasaurus polyzonus populations for various
correlations. First of all, I want to check if I there are differences in morphology
between the population, so that I would need to use distinct values for the different
populations}
\format{An object of class \code{character} of length 1.}
\usage{
path
}
\description{
General analysis of lizard data set
This is a script to check the data of Karusasaurus polyzonus populations for various
correlations. First of all, I want to check if I there are differences in morphology
between the population, so that I would need to use distinct values for the different
populations
}
\keyword{datasets}
