% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_boxplot.R
\name{m_boxplot}
\alias{m_boxplot}
\title{Function for simple boxplot
for statistical analysis}
\usage{
m_boxplot(x, y, data)
}
\arguments{
\item{x}{Variable to use for x axis}

\item{y}{Variable to use for y axis}

\item{data}{Data set to use}
}
\description{
Function for simple boxplot
for statistical analysis
}
