% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_boxplot.R
\name{m_boxplot}
\alias{m_boxplot}
\title{Function for simple boxplot
for statistical analysis}
\usage{
m_boxplot(x, y)
}
\arguments{
\item{x}{factor vector for catigorical variable on x axis}

\item{y}{vector for response variable on y axis}
}
\description{
Function for simple boxplot
for statistical analysis
}
