% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_microclim_input.R
\name{m_extract_microclim_input}
\alias{m_extract_microclim_input}
\title{Extract microclimate input
This function extracts the input data for the micro_global() function of NicheMapR
from our data structure (see example_lizard_data.csv in the testthat folder in tests)}
\usage{
m_extract_microclim_input(file, species = "")
}
\arguments{
\item{file}{Data file in csv format (set working directory to folder with all data files)}

\item{species}{Species of interest in format "Genus_species"}
}
\value{
List of vectors of micro climate data for each location
}
\description{
Extract microclimate input
This function extracts the input data for the micro_global() function of NicheMapR
from our data structure (see example_lizard_data.csv in the testthat folder in tests)
}
