% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_extract_microclim_input.R
\name{m_extract_microclim_input}
\alias{m_extract_microclim_input}
\title{Extract microclimate input}
\usage{
m_extract_microclim_input(location, loc_data)
}
\arguments{
\item{location}{An R factor containing the location ID of interest}

\item{loc_data}{A data frame with location IDs, coordinates and soil data}
}
\value{
Data frame of input data for given location for the micro_global function
}
\description{
This function extracts the input data of the micro_global() function of NicheMapR
based on our data structure (see example_coordinates.csv in the testthat folder in tests)
}
