% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_get_microclim.R
\name{m_get_microclim}
\alias{m_get_microclim}
\title{m_get_microclim}
\usage{
m_get_microclim(
  loc_row,
  nyears = 1,
  ndays = 12,
  timeper = "present",
  rcp = "none",
  shade = c(0, 90)
)
}
\arguments{
\item{loc_row}{A data frame row containing information on location
(location ID, latitude & longitude, nature of soil, soil reflectance,
soilgrids (downloading: yes or no))}

\item{nyears}{Number of years the model runs.}

\item{ndays}{Number of days modeled per year (12 - 365)}

\item{timeper}{Time period for climate data: either present or future ("present",
"2040_2059", "2080_2099")}

\item{rcp}{Pick a representative concentration pathway (rcp) for climate predictions
("none" (for present), "45" (medium low emissions), "85" (high emissions))}

\item{shade}{Vector of 2 numeric values: the first for minimum, the second for
maximum shade.}
}
\value{
list of locations (as sublists) with corresponding microclimate data
}
\description{
This function models the microclimate of locations in the given
input list (= output of m_extract_microclim_input)
}
