% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_all_micros.R
\name{m_plot_all_micros}
\alias{m_plot_all_micros}
\title{Plot all microclimates}
\usage{
m_plot_all_micros(multi_micro, save_plot = FALSE)
}
\arguments{
\item{multi_micro}{A list of climate scenarios which are lists of lists
of locations containing the 'metout' data structure of the output of the micro_global()
function (see vignette of microclimate in NicheMapR package)}

\item{save_plot}{Boolean whether the microclimate plot should be saved or not
(default = FALSE)}
}
\value{
Plot
}
\description{
This function plots the microclimate conditions from the output of
multiple micro_global models per location and scenario in one plot
}
