% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_ecto.R
\name{m_plot_ecto}
\alias{m_plot_ecto}
\title{Plot results of ectotherm function}
\usage{
m_plot_ecto(ecto, sim_name = ecto$LID, save_plot = F)
}
\arguments{
\item{ecto}{A list with the structure of the output of the ectotherm function (see
vignette of ectotherm in NicheMapR package).}

\item{sim_name}{A character string naming the simulation output.}

\item{save_plot}{Boolean whether the plot should be saved or not (default = FALSE).}
}
\value{
Plot
}
\description{
This function plots the output of the biophysical model of ectotherms
from the NicheMapR package.
}
