% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_ecto.R
\name{m_plot_ecto}
\alias{m_plot_ecto}
\title{Plot results of ectotherm function}
\usage{
m_plot_ecto(ecto, sim_name = ecto$LID, sub_title = "")
}
\arguments{
\item{ecto}{A list with the structure of the output of the ectotherm function (see
vignette of ectotherm in NicheMapR package).}

\item{sim_name}{A character string naming the simulation output.}

\item{sub_title}{A character string to be shown as subtitle of the plot.
This is to display some parameter values, like days and years the model is run for.}
}
\value{
Plot
}
\description{
This function plots the output of the biophysical model of ectotherms
from the NicheMapR package.
}
