% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_plot_micro.R
\name{m_plot_micro}
\alias{m_plot_micro}
\title{Plot microclimate conditions}
\usage{
m_plot_micro(multi_ecto)
}
\arguments{
\item{multi_ecto}{A list of lists with the structure of the output of the micro_global()
function (see vignette of microclimate in NicheMapR package).}
}
\value{
Plot
}
\description{
This function plots the microclimate conditions from the output of
multiple ectotherm models of different conditions (locations, times, emission
scenarios) together in one plot
}
