% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m_run_biophysical.R
\name{m_run_biophysical}
\alias{m_run_biophysical}
\title{m_run_biophysical}
\usage{
m_run_biophysical(
  liz_file,
  species,
  loc_file,
  loc_mean = FALSE,
  physio_file,
  nyears = 1,
  ndays = 12,
  burrow = FALSE,
  burrowtype = "sunshade",
  burrowdepth = 9,
  DEB = FALSE,
  timeper = "present",
  rcp = "none",
  shade = c(0, 90),
  save_plot = FALSE
)
}
\arguments{
\item{liz_file}{A file with lizard data (see example_lizard_data.csv) containing at least
one individual with location, morphology and reflectance data.}

\item{species}{The species to be looked at, which will be selected from the data frame and
split into sub groups of morphs if present.}

\item{loc_file}{File with information on locations (see example_coordinates.csv).}

\item{loc_mean}{Boolean whether the biophysical model should be run with location means
to decrease computational demand (loc_mean = TRUE) or if it should be run per individual (FALSE).}

\item{physio_file}{File path with physiological data of the lizards per location (see
'example_lizard_data.csv' for structure of dataframe).}

\item{nyears}{Number of years the model is run.}

\item{ndays}{Number of days modeled per year.}

\item{burrow}{Boolean whether lizard is allowed to seek shelter in burrow.}

\item{burrowtype}{Character string defining the burrow type: "sun" (always in the sun),
"shade" (always in the shade), "sunshade" (organism decides if burrow is in the sun
or in the shade).}

\item{burrowdepth}{The burrows depth in "soil nodes", a numeric ranging from 2 to 10.
2 equals to 2.5 cm, 3 to 5 cm, 4 to 10 cm, 5 to 15 cm, 6 to 20 cm, 7 to 30 cm, 8 to 50 cm,
9 to 100 cm and 10 to 200 cm. These are also the steps the lizard takes while burrowing.}

\item{DEB}{Boolean stating wheather the ectotherm should be run with or without
the Dynamic Energy Budget model.}

\item{timeper}{Character string of the time period for the climate data /
predictions ("present", "2040_2059", "2080_2099", "presentCCKP" to compare
the 1961-1990 of CCKP with the one from NicheMapR, and "presentNASA" to use more recent
(1983-2017) climate data).}

\item{rcp}{Character string to specify the emission scenario in case of climate
predictions ("none", "45", "85").}

\item{shade}{Vector of 2 numerics defining minimum and maximum shade.}

\item{save_plot}{Boolean whether plots shall be saved or not.}
}
\value{
List of lists with output of ectotherm function for each
species/population/entity
}
\description{
This function runs a full analysis of the climate niche starting with importing
our standard dataframe, running microclimate model and ectotherm function subsequently.
}
